Implement an explicit state machine for the main loop
This change cleans up some questionable code in App::ProcLaunch
that tried to keep track of old profiles, due to the bad design decision
to overwrite a list of profiles on every iteration of the main loop.
Instead, we keep track of a hash of profiles, updating it on every
loop if necessary. In addition, all of the code that decides what to do
when with regards to profiles has moved into App::ProcLaunch::Profile::run().

Also, add some docs about the changes.