Add a simple integration test (github issue #3)
This test is pretty simple but it exercises the core functionality.
It creates a simple profile like example_profiles/sleeper, waits
for it to start up, kills it, ensures that it restarts, and kills
the ProcLaunch instance.

This commit also changes log messages to be more consistent. Every
message is prefixed with either 'ProcLaunch' or the pertinent profile
name and messages that mention a pid always do so as 'pid <PID>'.