still_running will detect processes running as root
Perl's implementation of kill() returns the number of
processes that it was able to send a signal to. If that
number is 0, we have to check c's ERRNO, which in perl is
contained in the $! variable, already stringified.

The error will contain the message "Operation not permitted"
if the process is running but as a different user.